+++ common/data_refinery_common/models/gathered_accession.py
@@ -4,11 +4,11 @@ from django.db import models
 from django.utils import timezone
 
 
-class AccessionBacklogEntry(models.Model):
-    """Accession backlog entry model."""
+class GatheredAccession(models.Model):
+    """Gathered accession model."""
 
     class Meta:
-        db_table = "accession_backlog"
+        db_table = "gathered_accessions"
 
     code = models.TextField(unique=True)
     created_at = models.DateTimeField(auto_now_add=True)
@@ -21,7 +21,7 @@ class AccessionBacklogEntry(models.Model):
 
     def __eq__(self, other: object) -> bool:
         """Returns True if two objects are equal. Otherwise returns False."""
-        return isinstance(other, AccessionBacklogEntry) and self.code == other.code
+        return isinstance(other, GatheredAccession) and self.code == other.code
 
     def __hash__(self) -> int:
         """Returns accession object unique hash value."""
@@ -32,15 +32,15 @@ class AccessionBacklogEntry(models.Model):
         return ", ".join((self.code, self.technology, self.source, str(self.published_date.date())))
 
     @staticmethod
-    def create_from_ma_ae_entry(entry):
+    def create_from_ma_ae_entry(entry, organism=None):
         """Creates accession object from MicroArray ArrayExpress entry."""
-        accession = AccessionBacklogEntry()
+        accession = GatheredAccession()
         accession.code = entry["accession"]
         accession.source = "ebi_biostudies"
         accession.technology = "microarray"
 
-        if "organism" in entry:
-            accession.organism = entry["organism"]
+        if organism:
+            accession.organism = organism
         if "release_date" in entry:
             accession.published_date = timezone.make_aware(
                 datetime.strptime(entry["release_date"], "%Y-%m-%d")
@@ -51,7 +51,7 @@ class AccessionBacklogEntry(models.Model):
     @staticmethod
     def create_from_ma_geo_entry(entry):
         """Creates accession object from MicroArray GEO meta DB entry."""
-        accession = AccessionBacklogEntry()
+        accession = GatheredAccession()
         accession.code = entry["gse"]
         accession.source = "geo_meta_db"
         accession.technology = "microarray"
@@ -69,7 +69,7 @@ class AccessionBacklogEntry(models.Model):
     @staticmethod
     def create_from_rnaseq_entry(entry):
         """Creates accession object from RNA-Seq entry."""
-        accession = AccessionBacklogEntry()
+        accession = GatheredAccession()
         accession.code = entry["secondary_study_accession"]
         accession.source = "ebi_ena_portal"
         accession.technology = "rna-seq"
